+++ src/main/scala/uk/gov/hmrc/hmrcfrontend/views/Implicits.scala
@@ -18,9 +18,13 @@ package uk.gov.hmrc.hmrcfrontend.views
 
 import play.twirl.api.Html
 import uk.gov.hmrc.govukfrontend.views.html.components.implicits.{RichHtml => GovukRichHtml, RichOptionString => GovukRichOptionString, RichString => GovukRichString}
-import uk.gov.hmrc.hmrcfrontend.views.implicits.{RichCharacterCountSupport, RichDateInputSupport, RichErrorSummarySupport}
+import uk.gov.hmrc.hmrcfrontend.views.implicits.{RichAccountMenuSupport, RichCharacterCountSupport, RichDateInputSupport, RichErrorSummarySupport}
 
-trait Implicits extends RichCharacterCountSupport with RichErrorSummarySupport with RichDateInputSupport {
+trait Implicits
+    extends RichCharacterCountSupport
+    with RichErrorSummarySupport
+    with RichDateInputSupport
+    with RichAccountMenuSupport {
 
   implicit class RichHtml(html: Html) extends GovukRichHtml(html)